requirements.txt
streamlit==1.32.0
pandas==2.2.1
python-pptx==0.6.23
python-docx==1.1.0 